Fixed ccuc_id field type

+++ migrations/m140917_191401_init_2.php
@@ -127,7 +127,7 @@ class m140917_191401_init_2 extends CDbMigration
 
 
            CREATE TABLE `ccuc_user_company` (
-             `ccuc_id` tinyint(3) unsigned NOT NULL AUTO_INCREMENT,
+             `ccuc_id` int(3) unsigned NOT NULL AUTO_INCREMENT,
              `ccuc_ccmp_id` int(10) unsigned DEFAULT NULL,
              `ccuc_person_id` smallint(11) unsigned NOT NULL,
              `ccuc_status` enum('USER','HIDDED','PERSON','SYS') CHARACTER SET ascii NOT NULL DEFAULT 'PERSON',
